chore: move deployer logic into Deployer class
